api/requirements.txt
autopep8==1.4.4
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colormap==1.0.2
colourlovers==0.1.9
Flask==1.0.3
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-storage==1.16.1
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
Pillow==6.0.0
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
python-dotenv==0.10.3
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
Werkzeug==0.15.4
